Starting on wave function
Some problems with wave function.
Not made any significant progress so far, trying to figure out how to include classes to arrays as I am getting some errors.

My next step is adding an empty array with whether that area has been collapse, which options can be spawned and which mesh to generate there.
Unless I make the prototype handle it and cut out the options.