Changed approached
Decided to start working towards a small goal of using the generation I already have completed and simply do the first iteration of the WFC by simply changing the colour of one of the tiles generated. After this I'll work on propagating the new constraints to the adjacent tiles only.

For this I have decided to abandon some of the classes I started previously and keep most of the functionality inside a small amount of classes which I may later further divide.