Slot Behaviour WIP & Class delete
For now I will be using MyTestActor as the actor handling the slot behaviour and storing all the required information, outside the rules.

Added some of the functionality required to start working on the algorithm.

Fixed some of the problems with Enums from last update.

Started working on the required Structs. Will change the ModulePrototypes class to be a struct that contains all the relevant information for each tile.
Hopefully in the future I'll store all the tile information in a JSON file that gets loaded into a dictionary to be used for the rules/tile info.

Deleted some classes that won't be required.