Propagation working, While WIP
Managed to get the propagation working with 3 colours and simple rules.

After that I managed to get the WFC function working with the simple rules, however, it would collapse in 1 colour or 2 with mostly one colour taking over. After this I decided to try and expand the possible tiles to see if I could get different behaviour. This led to the engine freezing.

I tried to do some simple optimizations to cut down on the amount of for loops being ran by taking out the for loop being ran on every loop in HasCollapsed() function and replacing it with a variable containing a count of slots that had collapsed that gets updated every time one new slot is collapsed, and I ran into a crash error.

I'll take a break now, my next step is figuring out why it is currently freezing. To do this I may start working on a way to create the generation rules automatically with the use of "sockets" on each face of the tiles, comparing them with other tiles and writing down the possible neighbours on each direction. I would like to then export this as a JSON file that then gets used to load the rules.
If after this with the new rule set the function still freezes, I'll rule out the possibility of the rules generating the crash/freeze.

Will start refactoring my code soon.