Small Refactoring & Testing
Haven't had much time, so only some refactoring and testing the freezing problem.

Found out that with the 'collapseCount' variable, it wouldnt collapse all slots before matching the grid size, in the case of the current grid, 9 away from the total. However, this way the editor was not freezing, tried to swap back to the old method of looping with the collapseCount and it led to a freeze again.